+++ doctor-basic-api/src/main/java/io/terminus/doctor/basic/model/warehouseV2/DoctorWarehouseMaterialApplyPigGroupDetail.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.Builder;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -35,13 +36,13 @@ public class DoctorWarehouseMaterialApplyPigGroupDetail implements Serializable 
     private Integer type;
 
     //数量
-    private Double quantity;
+    private BigDecimal quantity;
 
     //单价
-    private Integer unitPrice;
+    private BigDecimal unitPrice;
 
     //金额
-    private Double amount;
+    private BigDecimal amount;
 
     //猪舍名
     private String pigBarnName;
